tszpufifo: Drain IRQ FD after opening
In zpu_fifo_init() the IRQ FD is opened. This may cause a
spurious IRQ to occur after its opened. After open()ing the FD,
seek to the begining and read a byte to "drain" it. This allows
future use of select() to not immediatly return on the first
access.