src/zpu_offload_demo: initial commit of userspace tools
The interface tool reads data out of the ZPU FIFO and pipes it
to the USB LCD screen attached. Handles connecting and
disconnecting the LCD gracefully. No FIFO communication happens
if no LCD connected.

The program tool is used to set up the attached LCD screen. This
only needs to be run once. It writes custom graphics and startup
message to NV storage on the LCD.